% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{as.eclairs}
\alias{as.eclairs}
\title{Convert SVD to eclairs}
\usage{
as.eclairs(
  dcmp,
  k,
  n,
  p,
  svd.method,
  mu = rep(0, p),
  sigma = rep(1, p),
  rn = NULL,
  cn = NULL,
  lambda = NULL
)
}
\arguments{
\item{dcmp}{SVD object}

\item{k}{rank of SVD}

\item{n}{number of samples}

\item{p}{number of features}

\item{svd.method}{method to compute SVD}

\item{mu}{per-feature means}

\item{sigma}{per-feature variances}

\item{rn}{rownames}

\item{cn}{column names}

\item{lambda}{shrinkage parameter. If not specified, it is estimated from the data.}
}
\description{
Convert SVD to eclairs object
}
