% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.mat.R
\name{autocorr.mat}
\alias{autocorr.mat}
\title{Create auto-correlation matrix}
\usage{
autocorr.mat(p, rho)
}
\arguments{
\item{p}{dimension of matrix}

\item{rho}{autocorrelation value}
}
\value{
auto-matrix of size p with parameter rho
}
\description{
Create auto-correlation matrix
}
\examples{
# Create 4x4 matrix with correlation between adjacent enties is 0.9
autocorr.mat(4, .9)

}
