% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate.R
\name{decorrelate_eclairs}
\alias{decorrelate_eclairs}
\alias{decorrelate_eclairs,matrix,eclairs-method}
\alias{decorrelate}
\alias{decorrelate,matrix,matrix,numeric,numeric-method}
\title{Decorrelation projection}
\usage{
decorrelate_eclairs(x, cor.est)

\S4method{decorrelate_eclairs}{matrix,eclairs}(x, cor.est)

decorrelate(x, U, dSq, sigSq)

\S4method{decorrelate}{matrix,matrix,numeric,numeric}(x, U, dSq, sigSq)
}
\arguments{
\item{x}{matrix to be transformed so *columns* are independent}

\item{cor.est}{estimate of correlation matrix from \link{eclairs} storing \code{U}, \code{dSq}, and \code{sigSq}}

\item{U}{orthonormal matrix with k columns representing the low rank component}

\item{dSq}{eigen values so that U diag(lamba) U^T is the low rank component}

\item{sigSq}{residual variance used for the scaled diagonal component: diag(sigSq, p)}
}
\description{
Given a vector \eqn{x ~ N(0, \Sigma)} where \eqn{\Sigma = U diag(\lambda) U^T + diag(\sigma^2)}, transform x so that it has an identity covariance matrix.  \eqn{\Sigma^{-0.5}x} is such a projection (see Strimmer whitening).  When \eqn{\Sigma} is \eqn{p \times p}, computing this project naively is \eqn{O(p^3)}.  Here we take advantage of the fact that \eqn{\Sigma} is the sum of a low rank decomposition, plus a diagonal matrix with a single value.
}
\details{
Decorrelation projection using low rank singular value decomposition of the correlation matrix
}
