% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_eclairs.R
\name{dmvnorm_eclairs}
\alias{dmvnorm_eclairs}
\alias{dmvt_eclairs}
\title{Density of the multivariate normal and t distributions}
\usage{
dmvnorm_eclairs(x, mu, ecl, log = FALSE)

dmvt_eclairs(x, mu, ecl, nu, log = FALSE)
}
\arguments{
\item{x}{numeric matrix with the data. The rows correspond to observations and the columns to variables.}

\item{mu}{mean vector}

\item{ecl}{covariance matrix as an \link{eclairs} object}

\item{nu}{degrees of freedom.  If finite, uses a multivariate t distribution}
}
\value{
A numerical vector with the density values calculated at each vector (row of the matrix x).
}
\description{
Density of the multivariate normal and t distributions
}
\examples{
library(Rfast)

X <- matrnorm(10000, 10)
mu <- colmeans(X)
Sigma <- cov(X)
ecl <- eclairs(X, lambda=1e-8)

# Multivariate normal
dmvnorm(X[1,], mu, Sigma, log=TRUE) 
dmvnorm_eclairs( X[1,], mu, ecl, log=TRUE)

# Multivariate Student t
dmvt(X[1,], mu, Sigma, nu=5, log=TRUE) 
dmvt_eclairs( X[1,], mu, ecl, nu=5,  log=TRUE)
}
\seealso{
\code{Rfast::dmvnorm()}, \code{eclairs()}
}
