% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_dependency.R
\name{effDependence}
\alias{effDependence}
\title{Evaluate effective dependence}
\usage{
effDependence(Sigma.eclairs)
}
\arguments{
\item{Sigma.eclairs}{estimate of correlation matrix from \code{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{lambda} and \eqn{nu}}
}
\description{
Evaluate effective dependence of the correlation matrix
}
\details{
Compute a metric of the average degree of dependence between all pairs of variables. Proposed by Pena and Rodriguez (2003), the 'effective dependence' is \eqn{1-|C|^\frac{1}{p}} where \eqn{C} is a correlation matrix between \eqn{p} variables. In the case where all pairwise correlations are \eqn{\rho}, the effective dependence converges to \eqn{\rho} for sufficiently large \eqn{p}.  Like the squared correlation, the value is bounded between 0 and 1.
}
\examples{
library(Rfast)
set.seed(1)
n = 2000 # number of samples
p = 1000 # number of features
rho = .9

# create correlation matrix
Sigma = matrix(rho, p, p)
diag(Sigma) = 1

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# eclairs decomposition
Sigma.eclairs = eclairs(Y, compute="cor")

# effective variance
effVariance( Sigma.eclairs )

# effective dependence
effDependence( Sigma.eclairs )

# expected value with constant correlation
rho

}
\references{
{
  \insertRef{pena2003descriptive}{decorrelate}
}
}
