% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_dependency.R
\name{effVariance}
\alias{effVariance}
\title{Evaluate effective variance}
\usage{
effVariance(Sigma.eclairs)
}
\arguments{
\item{Sigma.eclairs}{estimate of covariance/correlation matrix from \code{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{\lambda} and \eqn{\nu}}
}
\description{
Evaluate effective variance of the covariance/correlation matrix
}
\details{
Compute a metric of the amount of variation represented by a covariance/correlation matrix that is comparable across matrices of difference sizes. Proposed by Pena and Rodriguez (2003), the 'effective variance' is \eqn{|C|^\frac{1}{p}} where \eqn{C} is a correlation or covariance matrix between \eqn{p} variables. The effective variance is the mean of the log eigen-values.
}
\examples{
library(Rfast)
set.seed(1)
n = 2000 # number of samples
p = 1000 # number of features
rho = .9

# create correlation matrix
Sigma = matrix(rho, p, p)
diag(Sigma) = 1

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# eclairs decomposition
Sigma.eclairs = eclairs(Y, compute="cor")

# effective variance
effVariance( Sigma.eclairs )

}
\references{
{
  \insertRef{pena2003descriptive}{decorrelate}
}
}
