% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{ergpca}
\alias{ergpca}
\title{Empirically Reweighted Generalized PCA}
\usage{
ergpca(Y, k, lambda, k.features, cor.est = NULL, tol = 1e-05, maxit = 10000)
}
\description{
Empirically Reweighted Generalized PCA learns correlation structure between features in order to improve PCA projection
}
