% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_lambda.R
\name{estimate_lambda_eb}
\alias{estimate_lambda_eb}
\title{Estimate shrinkage parameter by empirical Bayes}
\usage{
estimate_lambda_eb(ev, n, p, nu)
}
\arguments{
\item{ev}{array of eigen values}

\item{n}{number of samples}

\item{p}{number of features}

\item{nu}{scale of prior covariance matrix}
}
\value{
value \eqn{\lambda} indicating the shrinkage between sample and prior covariance matrices.
}
\description{
Estimate shrinkage parameter by empirical Bayes
}
\details{
Estimate shrinkage parameter for covariance matrix estimation using empirical Bayes method (Leday and Richardson, 2019).  The shrinage estimate of the covariance matrix is \eqn{(1-\lambda)\hat\Sigma + \lambda \nu I}, where \eqn{\hat\Sigma} is the sample covariance matrix, given a value of \eqn{lambda}.  A large value of \eqn{\lambda} indicates more weight on the prior.
}
\examples{
ev = c(10, 2, 1) # eigen values
n = 12 # samples
p = 3 # features
nu = 2 # scale of target covariance

estimate_lambda_eb(ev, n, p, nu)

}
\references{
{
  \insertRef{leday2019fast}{decorrelate}
}
}
