% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca.R
\docType{class}
\name{fastcca-class}
\alias{fastcca-class}
\title{Class fastcca}
\description{
Class \code{fastcca}
}
\details{
Object storing:
\describe{
 \item{n.comp: }{number of canonical components}
 \item{cors: }{canonical correlations}
 \item{x.coefs: }{canonical coefficients for X}
 \item{x.vars: }{canonical variates for X}
 \item{y.coefs: }{canonical coefficients for Y}
 \item{y.vars: }{canonical variates for Y}
 \item{lambdas: }{shrinkage parameters from \code{eclairs}}
}
}
