% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca.R
\name{fastcca}
\alias{fastcca}
\title{Fast canonical correlation analysis}
\usage{
fastcca(
  X,
  Y,
  k = NULL,
  k.x = min(dim(X)),
  k.y = min(dim(Y)),
  lambda.x = NULL,
  lambda.y = NULL,
  svd.method = c("svd", "irlba", "pcaone")
)
}
\arguments{
\item{X}{first matrix (n x p1)}

\item{Y}{first matrix (n x p2)}

\item{k}{number of canonical components to return}

\item{k.x}{number of singular vectors of X to use}

\item{k.y}{number of singular vectors of Y to use}

\item{lambda.x}{shrinkage parameter for X. If \code{NULL}, estimate from data}

\item{lambda.y}{shrinkage parameter for Y. If \code{NULL}, estimate from data}

\item{svd.method}{SVD algorithm string "svd", "irlba", or "pcaone"}
}
\description{
Fast Canonical correlation analysis that is scalable to high dimensional data.  Uses covariance shrinkage and algorithmic speed ups to be linear time in p when p > n.
}
\examples{
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]

fastcca(pop, oec)

}
