% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca2.R
\name{fastcca2}
\alias{fastcca2}
\title{Fast canonical correlation analysis}
\usage{
fastcca2(
  X,
  Y,
  k = NULL,
  k.x = min(dim(X)),
  k.y = min(dim(Y)),
  lambda.x = NULL,
  lambda.y = NULL,
  svd.method = c("svd", "irlba", "pcaone")
)
}
\arguments{
\item{X}{first matrix (n x p1)}

\item{Y}{first matrix (n x p2)}

\item{k}{number of canonical components to return}

\item{lambda.x}{optional shrinkage parameter for estimating covariance of X. If NULL, estimate from data.}

\item{lambda.y}{optional shrinkage parameter for estimating covariance of Y. If NULL, estimate from data.}
}
\description{
Fast Canonical correlation analysis that is scalable to high dimensional data.  Uses covariance shrinkage and algorithmic speed ups to be linear time in p when p > n.
}
\examples{
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]

fastcca2(pop, oec)

}
