% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{getCov}
\alias{getCov}
\alias{getCor}
\alias{getCov,eclairs-method}
\alias{getCor,eclairs-method}
\title{Get full covariance/correlation matrix from \link{eclairs}}
\usage{
getCov(cor.est)

getCor(cor.est)

\S4method{getCov}{eclairs}(cor.est)

\S4method{getCor}{eclairs}(cor.est)
}
\arguments{
\item{cor.est}{eclairs decomposition}
}
\value{
p x p covariance/correlation matrix
}
\description{
Get full covariance/correlation matrix from \link{eclairs} decomposition
}
\details{
The full matrix is computationally expensive to compute and uses a lot of memory for large p.  So it is better to use \link{decorrelate} or \link{mult_eclairs} to perform projections in \eqn{O(np)} time.
}
