% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa,eclairs-method}
\alias{kappa,eclairs-method}
\title{Compute condition number}
\usage{
\S4method{kappa}{eclairs}(z, lambda = NULL)
}
\arguments{
\item{z}{\code{eclairs} decomposition}

\item{lambda}{specify lambda to override value from \code{z}}
}
\description{
Compute condition number of matrix from \code{eclairs} decomposition
}
