% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca.R
\name{predict,fastcca-method}
\alias{predict,fastcca-method}
\title{Predict one dataset from another}
\usage{
\S4method{predict}{fastcca}(object, X, Y, ...)
}
\arguments{
\item{object}{model fit from \code{fastcca()}}

\item{X}{first dataset}

\item{Y}{second dataset}

\item{...}{other arguments, not used}
}
\description{
Using CCA fit, predict one dataset from another
}
\details{
Specify which dataset to map from using \code{X} and \code{Y}.
}
\examples{
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]

# fit CCA
fit <- fastcca(pop, oec)

# predict Y given X
y.pred <- predict(fit, X = pop)
head(y.pred)

# predict X given Y
x.pred <- predict(fit, Y = oec)
head(x.pred)
}
