% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadForm.R
\name{quadForm}
\alias{quadForm}
\title{Evaluate quadratic form}
\usage{
quadForm(Sigma.eclairs, A)
}
\arguments{
\item{Sigma.eclairs}{estimate of covariance/correlation matrix from \code{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{\lambda} and \eqn{\nu}}

\item{A}{matrix}
}
\description{
Evaluate quadratic form
}
\details{
Evaluate quadratic form \eqn{A^T Sigma.eclairs^{-1} A}
}
\examples{
library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 200 # number of features

# create correlation matrix
Sigma = autocorr.mat(p, .9)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)

# eclairs decomposition
ecl = eclairs(Y)

# return scalar
quadForm( ecl, Y[1,])

# return matrix
quadForm( ecl, Y[1:2,])

}
