% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs_reform.R
\name{reform_decomp}
\alias{reform_decomp}
\title{Recompute eclairs after dropping features}
\usage{
reform_decomp(Sigma.eclairs, k = Sigma.eclairs$k, drop = NULL)
}
\arguments{
\item{Sigma.eclairs}{covariance/correlation matrix as an \link{eclairs} object}

\item{k}{the rank of the low rank component}

\item{drop}{array of variable names to drop.}
}
\value{
\link{eclairs} decomposition for a subset of features
}
\description{
Recompute eclairs after dropping features
}
\details{
Reform the dataset from the SVD, drop features, then recompute the SVD.  If the original SVD was truncated, then the reconstruction of the original data will be approximate.  Note that the target shrinkage matrix is the same as in \code{Sigma.eclairs}, so \eqn{\nu} is not recomputed from the retained features.
}
\examples{
library(Matrix)
library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 8*200 # number of features

# Create correlation matrix with autocorrelation
autocorr.mat <- function(p = 100, rho = 0.9) {
mat <- diag(p)
return(rho^abs(row(mat)-col(mat)))
}

# create correlation matrix
Sigma = autocorr.mat(p/8, .9)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# Correlation
#------------

# eclairs decomposition
Sigma.eclairs = eclairs(Y, compute="correlation")

# features to drop
drop = paste0("gene_",1:100)

# Compute SVD on subset of eclairs decomposition
ecl1 = reform_decomp( Sigma.eclairs, drop=drop)

ecl1

}
