% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_lambda.R
\name{series_start_total}
\alias{series_start_total}
\title{Create decreasing series of values}
\usage{
series_start_total(start, totalSum, n)
}
\arguments{
\item{start}{maximum value}

\item{totalSum}{specify target sum of values produces}

\item{n}{number of terms}
}
\description{
Create series of \eqn{n} values decreasing from \eqn{start} with a constant ratio between terms where the values sum to \eqn{totalSum}
}
