% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten.R
\name{whiten}
\alias{whiten}
\title{Decorrelation projection + eclairs}
\usage{
whiten(X, k = ncol(X), lambda = NULL)
}
\arguments{
\item{X}{matrix to be transformed so *columns* are independent}

\item{k}{the rank of the low rank component}

\item{lambda}{specify lambda and override value estimated by \code{eclairs}}
}
\description{
Efficient decorrelation projection using eclairs decomposition
}
\examples{

library(Matrix)
library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 8*200 # number of features

# Create correlation matrix with autocorrelation
autocorr.mat <- function(p = 100, rho = 0.9) {
mat <- diag(p)
return(rho^abs(row(mat)-col(mat)))
}

# create correlation matrix
Sigma = autocorr.mat(p/8, .9)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)

# eclairs decomposition
ecl = eclairs(Y)

# whitened Y
Y.transform = decorrelate(Y, ecl)

# Combine eclairs and decorrelate into one step
Y.transform2 = whiten(Y)

}
