% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.mat.R
\name{autocorr.mat}
\alias{autocorr.mat}
\title{Create autocorrelation matrix}
\usage{
autocorr.mat(p, rho)
}
\arguments{
\item{p}{dimension of matrix}

\item{rho}{autocorrelation value}
}
\description{
Create autocorrelation matrix
}
\examples{
# Create 10x10 matrix with correlation between adjacent enties is 0.9
autocorr.mat(4, .9)

}
