% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate.R
\name{decorrelate}
\alias{decorrelate}
\title{Decorrelation projection}
\usage{
decorrelate(X, cor.est)
}
\arguments{
\item{X}{matrix to be transformed so *columns* are independent}

\item{cor.est}{estimate of correlation matrix from \link{eclairs} storing \code{U}, \code{dSq}, and \code{lambda}}
}
\description{

}
\details{
Efficient decorrelation projection eclairs decomposition
}
