% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate.R
\name{decorrelate}
\alias{decorrelate}
\title{Decorrelation projection}
\usage{
decorrelate(X, Sigma.eclairs, lambda)
}
\arguments{
\item{X}{matrix to be transformed so *columns* are independent}

\item{Sigma.eclairs}{estimate of covariance/correlation matrix from \link{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{\lambda} and \eqn{\nu}}

\item{lambda}{specify lambda and override value from \code{Sigma.eclairs}}
}
\description{
Efficient decorrelation projection using \link{eclairs} decomposition
}
\details{
' FIX NOTATION HERE: Given a vector \eqn{x ~ N(0, \Sigma)} where \eqn{\Sigma = U diag(d_1^2) U^T + diag(\sigma^2)}, transform x so that it has an identity covariance matrix.  \eqn{\Sigma^{-0.5}x} is such a projection (see Strimmer whitening).  When \eqn{\Sigma} is \eqn{p \times p}, computing this projection naively is \eqn{O(p^3)}.  Here we take advantage of the fact that \eqn{\Sigma} is the sum of a low rank decomposition, plus a scaled identity matrix
}
