% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelate.R
\name{decorrelate}
\alias{decorrelate}
\title{Decorrelation projection}
\usage{
decorrelate(X, ecl, lambda, transpose = FALSE, alpha = -1/2)
}
\arguments{
\item{X}{matrix to be transformed so *columns* are independent}

\item{ecl}{estimate of covariance/correlation matrix from \link{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{\lambda} and \eqn{\nu}}

\item{lambda}{specify lambda and override value from \code{ecl}}

\item{transpose}{logical, (default FALSE) indicating if X should be transposed first}

\item{alpha}{default = -1/2.  Exponent of eigen-values}
}
\value{
a matrix following the decorrelation transformation
}
\description{
Efficient decorrelation projection using \link{eclairs} decomposition
}
\details{
' FIX NOTATION HERE: Given a vector \eqn{x ~ N(0, \Sigma)} where \eqn{\Sigma = U diag(d_1^2) U^T + diag(\sigma^2)}, transform x so that it has an identity covariance matrix.  \eqn{\Sigma^{-0.5}x} is such a projection (see Strimmer whitening).  When \eqn{\Sigma} is \eqn{p \times p}, computing this projection naively is \eqn{O(p^3)}.  Here we take advantage of the fact that \eqn{\Sigma} is the sum of a low rank decomposition, plus a scaled identity matrix
}
\examples{
library(Rfast)

n <- 800 # number of samples
p <- 200 # number of features

# create correlation matrix
Sigma <- autocorr.mat(p, .9)

# draw data from correlation matrix Sigma
Y <- rmvnorm(n, rep(0, p), sigma = Sigma * 5.1, seed = 1)

# eclairs decomposition
ecl <- eclairs(Y)

# whitened Y
Y.transform <- decorrelate(Y, ecl)

}
