% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca.R
\name{fastcca}
\alias{fastcca}
\title{Fast canonical correlation analysis}
\usage{
fastcca(
  X,
  Y,
  k = NULL,
  k.x = NULL,
  k.y = NULL,
  lambda.x = NULL,
  lambda.y = NULL,
  svd.method = c("svd", "irlba", "pcaone")
)
}
\arguments{
\item{X}{\code{matrix}, \code{data.frame}, \code{eclairs} decomposition or \code{list} from \code{svd()}}

\item{Y}{\code{matrix}, \code{data.frame}, \code{eclairs} decomposition or \code{list} from \code{svd()}}

\item{k}{number of canonical components to return}

\item{k.x}{number of singular vectors of X to use}

\item{k.y}{number of singular vectors of Y to use}

\item{lambda.x}{shrinkage parameter for X. If \code{NULL}, estimate from data}

\item{lambda.y}{shrinkage parameter for Y. If \code{NULL}, estimate from data}

\item{svd.method}{SVD algorithm string "svd", "irlba", or "pcaone"}
}
\description{
Fast Canonical correlation analysis that is scalable to high dimensional data.  Uses covariance shrinkage and algorithmic speed ups to be linear time in p when p > n.
}
\details{
Objects in \code{X} and \code{Y} are converted to \code{eclairs} decomposition from \code{matrix}, \code{data.frame} or \code{list} from \code{svd()}.  The \code{predict()} function maps from one input to another, except when input object is \code{list} from \code{svd()}.
}
\examples{
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]

# fit CCA
fit <- fastcca(pop, oec)
fit

# predict second dataset given first
y.pred <- predict(fit, X = pop)

# Advanced
# fastcca() also works with eclairs() objects
ecl.x <- eclairs(as.matrix(pop))
ecl.y <- eclairs(as.matrix(oec))
fit2 = fastcca(ecl.x, ecl.y)

# fastcca() also works mixed types
fit3 = fastcca(pop, ecl.y)

# it also works with results if svd()
}
