% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcca.R
\name{fastcca}
\alias{fastcca}
\title{Fast canonical correlation analysis}
\usage{
fastcca(X, Y, k = min(dim(X), dim(Y)), lambda.x = NULL, lambda.y = NULL)
}
\arguments{
\item{X}{first matrix (n x p1)}

\item{Y}{first matrix (n x p2)}

\item{k}{number of canonical components to return}

\item{lambda.x}{optional shrinkage parameter for estimating covariance of X. If NULL, estimate from data.}

\item{lambda.y}{optional shrinkage parameter for estimating covariance of Y. If NULL, estimate from data.}
}
\description{
Fast Canonical correlation analysis that is scalable to high dimensional data.  Uses covariance shrinkage and algorithmic speed ups to be linear time in p when p > n.
}
\details{
Results from standard CCA are based on the SVD of \eqn{Sig_{xx}^{-.5} Sig_{xy} Sig_{yy}^{-.5}}.

Uses eclairs and EB cov regularization, uses speed up of RCCA (Tuzhilina, et al, 2020) to perform CCA on n PCs and instead of p features.  Memory usage is n*p instead of p*p.  Computation is n^2p instead of p^3 of np^2
}
\examples{
pop <- LifeCycleSavings[, 2:3]
oec <- LifeCycleSavings[, -(2:3)]
fastcca(pop, oec)

}
\references{
{
  \insertRef{tuzhilina2020canonical}{decorrelate}
}
}
