% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{getCov}
\alias{getCov}
\alias{getCor}
\alias{getCov,eclairs-method}
\alias{getCor,eclairs-method}
\title{Get full covariance/correlation matrix from \link{eclairs}}
\usage{
getCov(Sigma.eclairs, ...)

getCor(Sigma.eclairs, ...)

\S4method{getCov}{eclairs}(Sigma.eclairs, ...)

\S4method{getCor}{eclairs}(Sigma.eclairs, ...)
}
\arguments{
\item{Sigma.eclairs}{eclairs decomposition}

\item{...}{other arguments}
}
\value{
p x p covariance/correlation matrix
}
\description{
Get full covariance/correlation matrix from \link{eclairs} decomposition
}
\details{
The full matrix is computationally expensive to compute and uses a lot of memory for large p.  So it is better to use \link{decorrelate} or \link{mult_eclairs} to perform projections in \eqn{O(np)} time.
}
