% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{getCov}
\alias{getCov}
\alias{getCor}
\alias{getCov,eclairs-method}
\alias{getCor,eclairs-method}
\title{Get full covariance/correlation matrix from \link{eclairs}}
\usage{
getCov(Sigma.eclairs, lambda, ...)

getCor(Sigma.eclairs, lambda, ...)

\S4method{getCov}{eclairs}(Sigma.eclairs, lambda, ...)

\S4method{getCor}{eclairs}(Sigma.eclairs, lambda, ...)
}
\arguments{
\item{Sigma.eclairs}{eclairs decomposition}

\item{lambda}{shrinkage parameter for the convex combination.}

\item{...}{other arguments}
}
\value{
p x p covariance/correlation matrix
}
\description{
Get full covariance/correlation matrix from \link{eclairs} decomposition
}
\details{
The full matrix is computationally expensive to compute and uses a lot of memory for large p.  So it is better to use \link{decorrelate} or \link{mult_eclairs} to perform projections in \eqn{O(np)} time.
}
\examples{
library(Matrix)
library(Rfast)
set.seed(1)
n = 800 # number of samples
p = 8*200 # number of features

# Create correlation matrix with autocorrelation
autocorr.mat <- function(p = 100, rho = 0.9) {
mat <- diag(p)
return(rho^abs(row(mat)-col(mat)))
}

# create correlation matrix
Sigma = autocorr.mat(p/8, .9)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)
Sigma = bdiag(Sigma, Sigma)

# draw data from correlation matrix Sigma
Y = rmvnorm(n, rep(0, p), sigma=Sigma*5.1)
rownames(Y) = paste0("sample_", 1:n)
colnames(Y) = paste0("gene_", 1:p)

# eclairs decomposition
Sigma.eclairs = eclairs(Y)

# extract covariance implied by eclairs decomposition
getCov(Sigma.eclairs)[1:3, 1:3]

}
