% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{getCov}
\alias{getCov}
\alias{getCor}
\title{Get full covariance/correlation matrix from eclairs}
\usage{
getCov(cor.est)

getCor(cor.est)
}
\arguments{
\item{cor.est}{eclairs decomposition}
}
\value{
p x p covariance/correlation matrix
}
\description{
Get full covariance/correlation matrix from eclairs decomposition
}
\details{
The full matrix is computationally expensive to compute and uses a lot of memory for large p.  So it is better to use \link{decorrelate} or \link{mult_eclairs} to perform operations in \eqn{O(np)} time.
}
