% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDet.R
\name{logDet}
\alias{logDet}
\title{Evaluate the log determinant}
\usage{
logDet(Sigma.eclairs, alpha = 1)
}
\arguments{
\item{Sigma.eclairs}{estimate of covariance/correlation matrix from \code{eclairs} storing \eqn{U}, \eqn{d_1^2}, \eqn{lambda} and \eqn{nu}}

\item{alpha}{exponent to be applied to eigen-values}
}
\description{
Evaluate the log determinant of the matrix
}
