% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{plot,eclairs-method}
\alias{plot,eclairs-method}
\title{Plot eclairs object}
\usage{
\S4method{plot}{eclairs}(x, y, ...)
}
\arguments{
\item{x}{eclairs object}

\item{y}{extra argument, not used}

\item{...}{additional arguments}
}
\value{
plot
}
\description{
Plot eclairs object
}
