% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{sv_threshold}
\alias{sv_threshold}
\title{Singular value thresholding}
\usage{
sv_threshold(n, p, d)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of features}

\item{d}{singular values}
}
\value{
Number of singular values to retian
}
\description{
Singular value thresholding evalues the number of singular values to retain
}
